Added comments to the updated_xarray_methods
For every new method, an empty dictionary is created called 'df_dict' which initially contains pandas dataframe for each variable in 'var_list'. The new function undergoes the method and adds the mean value of the current variable to it's corresponding  pandas dataframe. Then we close the xarray dataset and return the dictionaries containing pandas dataframe.